feat(api): Agregar endpoint de registro con logica de negocio

+++ src/MiApp.Application/Services/CategoryServices.cs
@@ -2,7 +2,6 @@ using MiApp.Application.Dtos;
 using MiApp.Application.Interfaces;
 using MiApp.Domain.Entities;
 using MiApp.Domain.Interfaces;
-
 namespace MiApp.Application.Services;
 
 /// <summary>
@@ -25,15 +24,16 @@ public class CategoryService : ICategoryService
         return categories.Select(c => MapToDto(c));
     }
 
-    public async Task<CategoryDto?> GetByIdAsync(int id)
+    public async Task<CategoryDto> GetByIdAsync(int id)
     {
         var category = await _repository.GetByIdAsync(id);
-        return category == null ? null : MapToDto(category);
+        if (category is null) throw new KeyNotFoundException($"No se encontró la categoría con el ID {id}");
+        return MapToDto(category);
     }
 
     public async Task<CategoryDto> CreateAsync(CreateCategoryDto createDto)
     {
-    
+
         var category = new Category(
             createDto.Name,
             createDto.Description
@@ -45,26 +45,28 @@ public class CategoryService : ICategoryService
 
     }
 
-    public async Task<CategoryDto?> UpdateAsync(int id, UpdateCategoryDto updateDto)
+    public async Task<CategoryDto> UpdateAsync(int id, UpdateCategoryDto updateDto)
     {
-      
+
         // 🔹 Buscar la entidad existente
         var existing = await _repository.GetByIdAsync(id);
-        if (existing == null) return null;
+        if (existing == null) throw new KeyNotFoundException($"No se encontró la categoría con el ID {id}");
 
         // 🔹 Actualizar los campos permitidos
         existing.Rename(updateDto.Name);
         existing.Description = updateDto.Description;
         existing.UpdatedAt = DateTime.UtcNow;
 
-         // 🔹 Guardar cambios usando el repositorio
-         var updated = await  _repository.UpdateAsync(existing);
-         return updated == null ? null : MapToDto(updated);
+        // 🔹 Guardar cambios usando el repositorio
+        await _repository.UpdateAsync(existing);
+        return MapToDto(existing);
     }
 
-    public async Task<bool> DeleteAsync(int id)
+    public async Task DeleteAsync(int id)
     {
-        return await _repository.DeleteAsync(id);
+        var category = await _repository.GetByIdAsync(id);
+        if (category is null) throw new KeyNotFoundException($"No se encontró la categoría con el ID {id}");
+        await _repository.DeleteAsync(id);
     }
